adapter: minor refactors around schema ID methods
This commit makes some minor refactors around the catalog methods for
accessing system schema IDs.

Most notably, it introduces a new constant `SYSTEM_SCHEMAS` that lists
the names of all system schemas and is then used to ensure that the
knowledge about the list of system schemas exists only in a single
place, rather than multiple that all need to be updated when the list
changes.

Furthermore, in a couple places `SchemaId` values are now passed by
value instead of by reference, simplifying the surrounding code a bit.
`SchemaId` is a `Copy` type, so there is no need to take values by
reference or explicitly clone them.