storage: deletion from differential collections
This commit extends the storage controller with support for deleting
contents of differential storage-managed collections. Instead of a list
of updates to append, the storage controller now accepts a
`StorageWriteOp` in its APIs to update these collections. In addition to
the existing append functionality, a `StorageWriteOp` can also specify a
deletion request based on a `Row` filter.

At the moment deletions are implemented simply by utilizing the
`desired` state that is kept in memory for differential collections,
making additional persist reads unnecessary. This will have to change
once the `CollectionManager` becomes smart enough to only keep the diff
between `desired` and the persisted collection contents.