catalog: Allow listening to durable changes
This commit adds the machinery necessary for the in-memory catalog to
listen to changes made to the durable catalog. Previously, the
in-memory catalog mocked this behavior via `Transaction`s. Currently,
this functionality is only used during bootstrapping.

When performing catalog transactions, the catalog applies individual
sub-transaction operations as they are processed. This is required due
to the fact that an operation my depend on the effects of the previous
operation. After the transaction commits, we drain all the durable
updates and throw them out since they were already applied.

An alternative approach would be to apply each sub-transaction
operation to a clone of the in-memory catalog. After the transaction
commits, we throw the clone out, listen to durable updates, and apply
those updates to the real in-memory catalog. We chose not to take this
approach because it would result in a non-trivial amount of duplicate
work.

Works towards resolving #24844