adapter: infrastructure for introspection subscribes
This commit introduces the overall infrastructure for introspection
subscribes, implemented by a new `coord::introspection` module that
extends the coordinator with methods to install introspection subscribes
and process updates for them.

This commit also introduces a first introspection subscribe that
surfaces dataflow error counts, as the first test case for this
infrastructure.