catalog: Plumb timestamps into state updates (#27778)
The in-memory catalog state updates are only processed for a single
timestamp at a time. Due to this, we previously omitted the timestamps
for convenience. Soon, we will want to be able to process updates for
multiple timestamps at once. In preparation for that, this commit
starts plumbing timestamps into the in-memory state updates.

Works towards resolving #24844