sql: Implement SHOW CREATE for managed clusters
This commit implements SHOW CREATE for managed clusters. It also gives
us a blueprint for how to convert catalog objects back to SQL
statements if we don't explicitly save the SQL statement. The process
looks roughly like:

1. Convert the catalog object to a plan.
2. Convert the plan to a statement.
3. Pretty print the statement to a string.

This process is designed to exactly mirror the process of converting
SQL to a catalog object.

Works towards resolving #15435