adapter: don't use max_query_result_size in RTW plans
Avoid max_query_result_size in the read portion of read-then-write plans.
Previously that session var would limit the size of the updates to its setting.
This would cause, for example, a `DELETE` issued from the console to fail if
there were more than 1MB of rows in a table.

Fixes #27428