storage: switch sources to columnated `DataflowError`
This change enables storage dataflows to make use of
columnated/lgalloced containers when exchanging data between operators.
We could have gone the other way and implemented a columnation region
for the `SourceReaderError` type but since there aren't any actual
downsides to switching to `DataflowError` we go for this simpler
approach.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>