adapter: read-only mode for tables and DDL
Work towards zero-downtime upgrades: https://github.com/MaterializeInc/materialize/issues/27406

The approach, while in read-only mode, is roughly:

- don't do things that we do periodically but don't carry content:
  periodic upper advancement
- stash changes to builtin tables and defer bootstrapping them to when
  we come out of read-only mode
- report an error back to the client when attempting things that require
  writing: table writes, DDL, things that would change the catalog

As of this change, we don't yet open the Catatalog in read-only mode,
and so we still apply migrations on bootstrap. We have to address that
for zero-downtime upgrades. And we don't yet make StorageCollections
aware of read-only mode, so we take over the critical SinceHandles.
Which we also still have to address.

As of this change, it is possible to bring up an environment in
read-only mode without error'ing/panic'ing. Previously, this would fail
because the storage controller was aware of read-only mode and would
error but components up the stack would not respect that.