adapter: read-only mode for tables and DDL
Work towards zero-downtime upgrades: https://github.com/MaterializeInc/materialize/issues/27406

The approach, while in read-only mode, is roughly:

- don't do things that we do periodically but don't carry semantic
  meaning:
  - periodic upper advancement
- stash changes to builtin tables and defer bootstrapping them to when
  we come out of read-only mode
- report an error back to the client when attempting things that require
  writing: table writes, DDL, things that would change the catalog

The code that allows us to stash changes and come out of read-only mode
gracefully (in process) is not strictly required because we plan on
using the halt!-then-restart (aka. crash-and-burn) approach for coming
out of read-only mode. However, the code was easy enough to add and
might come in handy in the future.

As of this change, we don't yet make StorageCollections aware of
read-only mode, so we take over the critical SinceHandles, which is also
a "write". This is addressed in a separate change.

As of this change, it is possible to bring up an environment in
read-only mode without error'ing/panic'ing. Previously, this would fail
because the storage controller was aware of read-only mode and would
error but components up the stack would not respect that.