bin/environmentd: auto-restart on halt
Bring the behavior of bin/environmentd more in line with Kubernetes and
auto-restart envd after it halts. This facilitates demos and local
testing of the new 0dt deployment flow, which involves a
halt-and-restart when the new generation takes over.

Kubernetes will also restart envd after a panic, but we choose not to do
so here, to avoid papering over panics during local development.

This commit involves a substantial refactoring and simplification of the
process group and session management in the run.py script. Moving the
subprocess into its own process group and session appears to be wholly
unnecessary; instead, we can simply create a new process group for the
parent, thus ensuring that all descendent processes are created in the
the process group for the run.py script, then send a SIGTERM to all
processes in that group when run.py exits.