ci/builder: pin to known good version of AWS CLI
The latest version (v2.17.2) has a weird glitch that causes the website
deploy job to occasionally fail to find EC2 instance role credentials.
Downgrade to an older version of the AWS CLI that was working properly.
Pinning the version in the CI builder is also just a healthy thing to
do, so this doesn't change on us unexpectedly the next time we have to
rebuild the CI builder image.