orchestrator: non-blocking `{ensure,drop}_service` calls
This commit makes the `ensure_service` and `drop_service` orchestrator
calls non-async and non-blocking. It achieves that by introducing an
orchestrator worker that runs in the background and asynchronously
handles requests for creating and removing services.

Both the k8s and the process orchestrator receive this treatment. We
mostly care about avoiding network calls to k8s on the coordinator's
main thread, but keeping the process orchestrator close to the k8s
orchestrator seems like a good idea.