ci/builder: upgrade to Hugo v0.128
The upgrade to Hugo v0.124.1 in #27679 has broken the deploy_website
build. The issue appears to be a bug in the Go AWS SDK version that Hugo
v0.124.1 uses that causes credential refreshing to sporadically fail.
The latest version of Hugo (v0.128) incoporates a Go AWS SDK upgrade
that hopefully fixes the issue.