compute: drop `CollectionState` eagerly
The dropping of collections happens in two phases, to ensure that
dataflow sources can stop reading from the inputs before we announce the
relinquishing of their read holds to the controller. Previously this was
done by keeping the `CollectionState` alive until
`report_dropped_collections`, so that method could inspect the reported
frontiers. Keeping the `CollectionState` alive also meant keeping its
`CollectionLogging` alive.

This made it possible for two `CollectionLogging` instances to be alive
for the same collection after reconciliation. When reconciliation drops
and reinstalls a collection, it makes sure that the `AllowCompaction`
command is handled before the `CreateDataflow` command. Handling
`AllowCompaction` results in a `drop_collection` call, but intentionally
not in a `report_dropped_collections` call (that is deferred until after
the next Timely step), so the old `CollectionLogging` instance stays
alive. Then handling `CreateDataflow` creates a new `CollectionLogging`
instance for the same collection ID. Creation and deletion of
`CollectionLogging` instances causes emission of log events, and the
logging dataflows assume that collection IDs are unique, so having two
`CollectionLogging` instances for the same ID alive corrupts the logging
dataflow state.

The fix is to make sure `drop_collection` also drops the
`CollectionLogging` instance of the collection, as it already does with
sink tokens. This commit does that, but to prevent issues like this from
reappearing when we add more fields to `CollectionState`, it also
introduces a separate `DroppedCollection` type that only holds the state
required by `report_dropped_collections`. This forces us to be explicit
about which fields we want to survive `drop_collection` and lets us make
sure that none of them holds onto any shared resources.