Claim compactions when all merges are empty
If a shard has a series of empty batches appended to it, it will
generate a new empty-batch compaction for every append.  This can
prevent our new claim-unclaimed mechanism from kicking in, since it will
only trigger extra compactions when the list of requested compactions is
empty.

The fundamental behaviour here is a bit questionable, but let's say that
we're willing to tack on new compactions whenever all the input batches
are empty / thus don't actually require any "real" compaction work.