bin/environmentd: fix Ctrl-C in noninteractive shells
The changes in 035b7dd resulted in Ctrl-C no longer working when
bin/environmentd was wrapped in a shell script.

The issue was that, when bin/environmentd is run from a noninteractive
shell[^1], the shell won't spawn a dedicated process group for the
wrapper Python script, and so calling `os.setpgid` would cause the
process to no longer be the foreground process group, which means it
would no longer receive signals like SIGINT.

The fix is to bring back a piece of the process management approach from
before 035b7dd. We fork a *new* process to run the child in and assign
*that* process a dedicated process group. The parent bin/environmentd
process remains in the foreground process group and thus receives
signals, and so can tell the child process and all descendent processes
to exit when the parent receives a signal and decides to exit.

[^1] Technically a shell with job control disabled, which could be an
interactive shell,  but in general interactive shells have job control
enabled and noninteractive shells have job control disabled.