storage: in StorageCollections, add a facade for SinceHandle
Work towards zero-downtime upgrades:
https://github.com/MaterializeInc/materialize/issues/27406.

We want to be able to bring up a `StorageCollections` in read-only mode.
In that mode, it cannot acquire and drive forward critical
`SinceHandles`. The idea is to be able to work with either a leased
handle or a critical handle, depending on the read-only bit. This puts a
facade in place that will make it so the code that actually uses the
handle doesn't have to be aware of the difference.