deps: upgrade to unpatched librdkafka
To work around an unexplained memory error involving OpenSSL, we had to
revert a librdkafka commit that changes how Kafka clients load CA
certificates (f8830a2).

However, after extensive review, neither @petrosagg nor I can spot a
memory error in that commit. That suggests that there might be a
preexisting memory error in either OpenSSL or librdkafka that f8830a2
exacerbates. That's scary, because there might be a way to trigger that
memory error even with f8830a2 reverted.

This commit upgrades us to mainline librdkafka v2.4.0--i.e., a version
of librdkafka with f8830a2 unreverted--to make the memory error
reproducible. That will allow us to debug and find the root cause of the
memory error.

FOR DEBUGGING PURPOSES ONLY. Do NOT merge this commit in its current
state.