deps: upgrade to a rdkafka with DNS resolution callback
The last piece of the puzzle for denying connections to local IPs
(MaterializeInc/console#1635) is to intercept DNS calls made by
librdkafka and reject any DNS responses that contain local IPs.

This commit upgrades to a version of rdkafka that permits overriding the
DNS resolution callback in full. The previous version permitted
rewriting the hostname and port prior to DNS resolution, but not
inspecting the resolved addresses.

Actually wiring up the DNS resolution callback to use the
`ore::net::lookup_host` function that rejects local IPs is left to a
future commit to ease the review of this commit. There should be no
behavior changes as a result of this commit.