adapter: unified hydration time introspection
This commit adds a unified introspection relation exposing the hydration
time of compute collections, `mz_compute_hydration_times`. This is a
"retained metrics relation" that has a compaction window of 30 days,
making it possible to get at historical hydration times. To ensure that
the storage size of the relation doesn't become excessive, we filter out
non-transient dataflows and only include hydration times of indexes and
materialized views.