compute: mz_compute_hydration_statuses view
This commit changes `mz_compute_hydration_statuses` from a
storage-managed collection to a view that is defined in terms of
`mz_compute_hydration_times`, and also adds the hydration time as an
additional column. This allows us to rip out some of the hydration
status tracking from the compute controller.

With this change `mz_compute_hydration_statuses` receives some
properties of `mz_compute_hydration_times`, namely the exclusion of
transient collections and the retained history. On the other hand, the
view query ensures that MVs that have advanced to the empty frontier are
still included in `mz_compute_hydration_statuses` even though they don't
appear in `mz_compute_hydration_times`. This is done to ensure our
recommented blue-green readiness query keeps working.