catalog: Fix formatting in initialize_state (#27978)
Previously, the function body of `initialize_state` was wrapped in an
async block. For some reason this block broke `rustfmt` and prevented
automatic formatting. The async block was used to box the returned
future since it was very large, and we didn't want it on the stack.

This commit removes the async block and instead updates callers of
`initialize_state` to box the returned future which fixes formatting. In
release mode, the optimizer *should* create the future in the heap
instead of creating it on the stack and moving it to the heap.

The non-whitespace characters of `initialize_state` are completely
unchanged, however the indentation has been changed which makes the diff
large.