catalog: Add multi-subscriber catalog
This commit adds the functionality for a read-only catalog to run
concurrently with a writer catalog, and have the read-only catalog
subscribe to changes made by the writer. The state of read-only catalog
is kept up to date with the durable catalog via the changes.

This functionality is only used in tests currently.

Works towards resolving #24844