adapter: always declare MV imports non-monotonic
Every materialized view has at its output a self-correcting persist sink
that reserves the right to insert "corrections" whenever it notices that
its output collection differs from the desired state. Across releases
this can happen (e.g. when we fix a bug in the optimizer or rendering),
so we always need to assume that a persist sink can emit retractions
into the MV collection. This makes all MVs non-monotonic, and treating
them as monotonic would be a mistake.