catalog: Reorganize mutable CatalogState
This commit re-organizes the fields of `CatalogState` into two
sections. State that is derived from the durable catalog and state that
isn't derived from the durable catalog. Additionally, all mutations to
state that is derived from the durable catalog is moved to either
`open.rs` or `apply.rs`.

This commit is 100% code movement. The goal is to increase the
confidence that the in-memory catalog state that is derived from the
durable catalog is only ever updated by applying updates from the
durable catalog.

Works towards resolving #24844