Switch to Polar Signals-maintained jemalloc_pprof
Discussed with Frederic and me (from PS)
with various Materialize folks when we visited the office last week:
After Materialize donated this code to Polar Signals, they never
updated to the new upstream. It is better to do so, because then
Materialize can easily pick up any new features or bugfixes the PS team works on.