build: fix disabling jemalloc in sanitizer builds
CI was previously building with the `--workspace` flag, which force
includes all crates in the workspace. In particular, this includes the
`mz-alloc-default` crate, which force enables jemalloc, even if
`--no-default-features` was passed with the intent to disable jemalloc.

To work around this behavior, this commit changes the definition of
`default-members` to include all crates *except* for the two crates that
cause problems: mz-alloc-default and mz-fivetran-destination. That means
that CI and end users no longer need to pass `--workspace`, because
exactly the right crates are included by default. (Folks who want to
build the Fivetran destination are already aware that they need to
specify `-p mz-fivetran-destination`.)

Since maintaining `default-members` by hand is impractical, this commit
adds a new Cargo lint which ensures that `default-members` looks as it
should.