storage/source/kafka: downgrade based on last_offsets, not librdkafka
This commit fixes a flake observed in the source-sink-errors test
(#28016).

Here's what was happening:

  * The source would read the message with offset 1 from Kafka.
  * The source would emit the message to its data output.
  * The test would turn Redpanda off.
  * librdkafka would notice the broken connections and report the
    offsets for all partitions as "invalid".
  * The source would fail to downgrade its capability to offset 2
    because librdkafka was no longer reporting a valid offset.
  * The test would turn Redpanda on.
  * librdkafka would successfully reconnect but not update the offset
    for the partition.
  * The source would not downgrade its capability to offset 2 because
    librdkafka was still not reporting a valid offset.

If a new message was written to Kafka at offset 2, then librdkafka
*would* emit the message and report a valid offset of 3, at which point
the source would downgrade its capability to offset 3, and both the
message at offset 1 and the message at offset 2 would become visible.

The flake started occurring after the upgrade to librdkafka v2.4. It's
not clear whether there was a behavior change inside of librdkafka, or
whether its simply a timing condition that always existed and somehow
became more common in the new version.

Regardless, the fix is straightforward: attempt to seek the partition
client whenever an invalid offset is observed. This will prevent the
source from getting stuck when such an ill-timed Kafka reconnection
causes offsets to become invalid.

Fix #28016.