storage/source/kafka: don't seek
This commit fixes a flake observed in the source-sink-errors test
(#28016).

Here's what was happening:

  * The source would read the message with offset 1 from Kafka.
  * The source would emit the message to its data output.
  * The test would turn Redpanda off.
  * librdkafka would notice the broken connection and, incorrectly,
    redeliver the message with offset 1. (This is a known problem with
    librdkafka, and something the code already protects against.)
  * The source would seek to offset 2, thus marking its position within
    librdkafka as "invalid".
  * The source would fail to downgrade its capability to offset 2
    because librdkafka was no longer reporting a valid offset.
  * The test would turn Redpanda on.
  * librdkafka would successfully reconnect but not update the offset
    for the partition.
  * The source would not downgrade its capability to offset 2 because
    librdkafka was still not reporting a valid offset.

If a new message was written to Kafka at offset 2, then librdkafka
*would* emit the message and report a valid offset of 3, at which point
the source would downgrade its capability to offset 3, and both the
message at offset 1 and the message at offset 2 would become visible.

The flake started occurring after the upgrade to librdkafka v2.4. It's
not clear whether there was a behavior change inside of librdkafka, or
whether its simply a timing condition that always existed and somehow
became more common in the new version.

Regardless, the fix is straightforward: don't attempt to seek the
partition. librdkafka has only been observed to redeliver a single
message at a time, so there is no need to seek forward to skip messages.
We can simply receive and discard the errant messages normally.

Fix #28016.