catalog: Fix savepoint catalog listen behavior
Savepoint catalogs are able to accept writes without erroring, but the
writes are never committed durably. The main usage of savepoint
catalogs is to validate an upgrade in-memory before actually executing
the upgrade.

Previously, savepoint catalogs would update their in-memory state
based on the savepoint transactions committed by themselves AND by
listening to durable changes made by any writer catalog. This is NOT
correct because there is no attempt to check for conflicts or reconcile
the writes made by the two separate catalogs.

This commit fixes the issue by ignoring all durable catalog updates in
savepoint catalogs after the initial snapshot. A correct mental model
for this is that creating a savepoint catalog is like creating a new
branch of catalog state. This branch does not affect the main catalog
state branch, nor does it see changes from the main branch.

Currently, there is no way to merge this savepoint branch back into the
main catalog state branch. In order to achieve zero-downtime upgrades,
we will need to figure out a way to merge these two branches.