storage/sink/kafka: record number of consumed progress records
We want a metric that reveals whether compaction is working on users'
progress topics.

The outstanding_progress_records metric is close, but the low water mark
of a topic may get stuck arbitrarily far in the past if a sink is
dropped, so the delta between a progress topic's high and low water
marks is not a good measure of whether a user has enabled compaction.

Instead, introduce a consumed_progress_records metric, which directly
records the number of messages in a progress topic for each sink. If
this number ever ticks up into the hundreds of thousands or millions,
that is a clear sign that compaction is not working.