test: attempt to deflake source-statistics.td
Because the environment is not reset between testdrive runs, and we are
querying retained-metrics collections with `AS OF AT LEAST 0`, there is
a risk that we are seeing contents from previously run testdrive files.
Avoid this possibility by using a unique name for the load generator
source we filter on.