catalog: Fixup savepoint catalog commit and sync (#28114)
This commit is a followup to cef629e91d81eab91a7f200e7e72c056d8fabe27
that adds the following fixup features:

- Commits in savepoint catalogs increment the in-memory catalog upper.
This isn't strictly necessary in savepoint mode since the upper isn't
used for anything, however it matches the behavior of write mode.
- Sync sets the current upper to the max of the target upper and the
current upper, instead of blindly setting it to the target upper. This
protects against a caller trying to sync in the past and causing the
upper to decrease, which should be impossible. There is no current code
path where a caller would pass a timestamp less than the current upper,
but it's better to protect against it.