storage: Prevent finalization in read-only mode
This commit prevents shard finalization in read-only mode. Read-only
mode should not affect changes to external systems. Previously, if a
shard was present in the unfinalized/finalizable shard list on a read
only instance, which could happen as a result of a well-timed startup,
the read-only instance would attempt to forcibly downgrade the shard's
upper and since and finalize the shard.

Works towards resolving #27981