adapter: serialize DDL
Serialize most DDL by queueing DDL while another is executing. Use a locking
VecDeque helper (and refactor another use of that to use the new helper) to
manage state.

The lock is either acquired for the required statements and stored in the conn
meta or the statement is enqueued. When the transaction is complete the lock is
dropped and the coordinator is sent a message to run the next queued statement.

Fixes #27821