adapter: enforce a lower bound for max_result_size
This commit enforces a lower bound of 1MB for the `max_result_size`
parameter. This parameter is meant to prevent environmentd OOMs and
should therefore be set to a value proportional to environmentd's
available memory size. It's save to assume that we'll never try to run
environmentd with available memory in the order of megabytes.

Enforcing a lower bound ensures that nobody can accidentially set
`max_result_size` to a value that is small enough to interfer with
system queries, like introspection subscribes or those performed by our
metrics infrastructure.