persist: Columnar Encoding V2 (#27961)
This PR adds a "V2" of columnar encoding to Persist, behind the dyncfg
`persist_batch_columnar_format` with the value `both_v2`.

Note: today we have a "V1" columnar encoding which encodes non-trivial
types (e.g. Array or JSON) into its `ProtoDatum` representation. This PR
uses the new encodings introduced in
https://github.com/MaterializeInc/materialize/pull/27084 which don't
rely on `ProtoDatum` at all.

### Changes

#### Statistics
These changes are in commit 2.

With the new columnar encodings, the types we persist for some
statistics have changed, e.g. for columns of type `Timestamp` we collect
`PrimitiveStats<i64>` instead of `AtomicByteStats`. To handle this
change we introduce a new method on
[`RelationPartStats`](https://github.com/MaterializeInc/materialize/blob/main/src/storage-types/src/stats.rs#L34),
`fn col_values2(...)` which mirrors the existing `fn col_values(...)`.
The new method knows how to interpret stats from both the existing
types, and the new types.

When getting a `ResultSpec` for a column we call both `col_values(...)`
and `col_values2(...)`, `soft_assert` that they match, and on the off
chance that they don't, return the value from `col_values(...)` to match
today's behavior. Calling both methods allows us to do a slow rollout of
the new encoding types, while getting validation that the new code
exhibits the same behavior as the existing code.

I also added a `proptest` that encodes updates with `PartBuilder` and
`PartBuilder2`, then calls both `col_values(...)` and `col_values2(...)`
on both generated parts, and asserts the returned `ResultSpec`s match.

Why not add a version flag to `PartStats`? An alternative implementation
would add a version field to `PartStats`, and interpret the statistics
based on the version. I tried to do this, but couldn't figure out how to
add the version field, without requiring a large migration of our
protobuf types.

#### Parquet Encoding
These changes are in commit 3.

To simplify things we only encode and decode structured data at the
Parquet level with the new columnar version. If we find a `Part` with
metadata that indicates it was encoded with `V1`, we ignore the `k_s`
and `v_s` columns for that `Part`. We also change the Parquet encoding
for a Row to use the _column indexes_ as Parquet field names, instead of
the column names from the `RelationDesc`, this was the naming scheme
specified in the design doc.

What's a little tricky is we still use column names in the statistics.
Doing so reduces the size of this PR and allows us to validate the new
stats code with the Row encoded data and the Columnar encoded data.
Encoding column indexes with statistics is an incremental change we can
make in the future.

#### Updating the `persist_batch_columnar_format` dyncfg
These changes are in commit 4.

This PR also updates the recently added `both_v1` variant of the
`persist_batch_columnar_format` to `both_v2`. Originally I added
`both_v1` as the variant for the _new_ encoding, and considered the
current encoding V0. But later I realized in the [Parquet metadata we
encode the current version as
V1](https://github.com/MaterializeInc/materialize/blob/main/src/persist/src/indexed/encoding.rs#L467).
It's not super important, but I figured it would be easier to reason
about versions if the metadata in the Parquet encoding matched up with
the dyncfg and what we [write in
State](https://github.com/MaterializeInc/materialize/blob/main/src/persist-client/src/batch.rs#L1134).

#### `PackedNaiveDateTime`
These changes are in commit 5.

This PR also adds a new `FixedSizeCodec` implementation which is used
for `Timestamp` and `TimestampTz`. Previously we were encoding
timestamps as microseconds since the epoch in an `i64`. This is what
Postgres does and covers the entire range of the `Timestamp` type. This
doesn't work though because we allow ingesting timestamps with
nanosecond precision from Kafka, and our current durable representation
(`ProtoRow`) can roundtrip this.

To prevent making our Persist encoding lossy, we now encode timestamps
in a format that more directly reflects the internal `chrono::DateTime`
used to represent them. Benchmarks show the packed encoding supports ~3x
higher throughput than the previous conversion to microseconds.

### Motivation

Progress towards:
https://github.com/MaterializeInc/materialize/issues/24830

### Tips for reviewer

There is a bit of complexity in this PR, it's broken into separate
commits which hopefully makes it easier to review.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly
considered. ([trigger-ci for additional test/nightly
runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design
doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md),
is a design doc
([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)),
or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T`
mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md)
(possibly in a backwards-incompatible way), then it is tagged with a
`T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests,
there is a companion cloud PR to account for those changes that is
tagged with the release-blocker label
([example](https://github.com/MaterializeInc/cloud/pull/5021)).
<!-- Ask in #team-cloud on Slack if you need help preparing the cloud
PR. -->
- [x] This PR includes the following [user-facing behavior
changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - N/a