storage: Remove invalid assert from startup (#28181)
One of the responsibilities of StorageCollectionsImpl::initialize_state
is to take all the global IDs known by the adapter catalog and ensure
that it is included in the storage metadata. Previously, there was an
assert that asserted that all IDs known by the catalog but not known by
storage MUST be system objects. The likely intuition of this assert is
that catalog data and storage metadata are always updated atomically, so
any IDs not know by storage must be new system objects added as part of
a version upgrade. However, this is not correct. Builtin item migrations
may also create new user objects (after dropping the old ones).

This commit removes the assert, which is not always correct as described
above. Investigation into the surrounding code seems to indicate that
this assertion was not needed for correctness, but was a sanity check.

Previous discussions of this assert have led to discussions about the
fact that if any of the new user IDs are sinks, then the sink will
produce duplicate data. This is a known issue, #18767. This commit adds
an error log for this specific scenario.

Additionally, while we're touching `initialize_state`, this commit
updates a set difference calculation to use the builtin `difference`
function.

Finally, we update the builtin migration test to trigger a migration of
all builtin tables AND views instead of just views, and the test now
creates a user object that will be migrated. This stresses more of the
builtin migration framework than was previously tested, but
unfortunately would not have triggered the assert since the controller
is excluded from the test.