catalog: Optimize memory of durable catalog
Previously, the durable catalog stored two copies of every catalog
update. One copy is used to create durable transactions, while the
other copy is used to serve catalog updates to subscribers. This commit
updates the durable catalog so that the same copy of each update can be
used to both create new transactions and serve subscribers.

The basic idea is that the durable catalog stores a vector of
(update, time, diff) tuples along with an upper indicating times that
haven't been consumed by the subscriber. Only times less than the
subscriber can be consolidated, and only times greater than or equal to
the subscriber upper are served to subscribers.