catalog: Fix off by one error in sync_updates
Previously, the `sync_updates` function in the durable catalog claimed
to return all updates up to and including the provided timestamp. In
reality, it was only returning all updates up to (not including) the
provided timestamp.

This commit fixes the documentation of `sync_updates` and all callers.
This function only has a single caller, which uses its output for a
soft_assert. Fixing the function itself to include the timestamp would
be much more useful, but much harder to implement. Since it's not
heavily used, we opt for the easy route.