catalog: Reduce memory usage during startup (#28208)
Previously, we would keep two copies of each storage usage event in
memory during startup. One copy was used to generate builtin table
updates while the other copy was used to prune old storage usage events.
In some environments, storage usage events are so large, that this
duplication had significant memory impacts during startup.

This commit fixes the issue by condensing the two copies into a single
copy that is used for both pruning and generating builtin table updates.

As a consequence, catalogs will not receive any storage usage event
updates while subscribing to the durable catalog. For now this is fine,
because we only use the subscription during startup, which special cases
storage usage events. In the future we'll likely have to resolve this
somehow.

The duplication was introduced by

https://github.com/jkosh44/materialize/commit/b33485487bf439ac2e4cd2557fdaea0cd043da12,
so in spirit this commit is a
partial revert of
https://github.com/jkosh44/materialize/commit/b33485487bf439ac2e4cd2557fdaea0cd043da12.