storage/upsert: fix `select!`/`Stream` bug
When selecting between multiple async streams it's important that a
`Some(_)` pattern is used so that when one of the streams finishes and
yields `None` that branch of `select!` is marked as disabled and another
one is waited for.

This problematic pattern in `upsert_inner` could trigger a busy loop
where when the `input` stream finished it would continously yield
`None`, which prevented the thread from running anything else and
eventually existing the loop.

Fixes #27835

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>